+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add fabric SPAN source group
 - Include default values in module
 - BREAKING CHANGE: `depends_on` can no longer be used to express explicit dependencies between NaC modules. The variable `dependencies` and the output `critical_resources_done` can be used instead, to ensure a certain order of operations.
+- Add config passphrase
 
 ## 0.3.4